PM_TIME_SEC into 64bit

+++ lib/libc/sys-minix/stime.c
@@ -14,6 +14,6 @@ int stime(time_t *top)
   message m;
 
   memset(&m, 0, sizeof(m));
-  m.PM_TIME_SEC = (long)*top;
+  m.PM_TIME_SEC = *top;
   return(_syscall(PM_PROC_NR, PM_STIME, &m));
 }
